Change bio Int for Double and setup bio info on screen

+++ PokeWiki/Sources/Model/PokemonBasicInfo.swift
@@ -14,6 +14,6 @@ struct PokemonBasicInfo {
 }
 
 struct PokemonBioInfo {
-    let height: Int
-    let weight: Int
+    let height: Double
+    let weight: Double
 }
